+++ crates/config/src/workspace/experiments_config.rs
@@ -6,9 +6,11 @@ config_struct!(
     #[derive(Config)]
     pub struct ExperimentsConfig {
         /// Track and determine affected projects and tasks asynchronously.
+        #[setting(env = "MOON_EXPERIMENT_ASYNC_AFFECTED_TRACKING")]
         pub async_affected_tracking: bool,
 
         /// Build the project and task graphs asynchronously.
+        #[setting(env = "MOON_EXPERIMENT_ASYNC_GRAPH_BUILDING")]
         pub async_graph_building: bool,
     }
 );